+++ packages/terra-core-docs/src/terra-dev-site/test/text/TextVisuallyHidden.test.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import Text from 'terra-text';
+
+const TextVisuallyHidden = () => (
+  <div>
+    <span>Text below is visually hidden but available to screen readers</span>
+    <Text id="textVisuallyHidden" isVisuallyHidden>Visually Hidden</Text>
+  </div>
+);
+
+export default TextVisuallyHidden;